hacker_news/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
tomli==1.0.4
